@@ -15,7 +15,7 @@
  */
 package com.example.android.miwok;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 public class NumbersActivity extends AppCompatActivity {
