@@ -1,11 +1,9 @@
 package com.example.android.miwok;
 
 import android.app.Activity;
-import android.content.Context;
-import android.media.Image;
-import android.media.MediaPlayer;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
